ADD: one to many and many to many

+++ laravel-boolpress/app/Http/Controllers/Admin/PostController.php
@@ -9,6 +9,7 @@ use App\Models\Category;
 use App\Models\Post;
 use App\Models\Tag;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Str;
 use Illuminate\Validation\Rule;
 
@@ -53,6 +54,11 @@ class PostController extends Controller
         // dd($request->all());
         // dd($request->all(), $request->validated());
 
+        // Recuperiamo l'id dell'utente loggato
+        // $user_id = Auth::id();
+        // $user_name = Auth::user()->name;
+        // dd(Auth::user());
+
         $form_data = $request->validated();
 
         // dd($form_data);
@@ -73,6 +79,7 @@ class PostController extends Controller
         } while ($find !== null);
 
         $form_data['slug'] = $slug;
+        $form_data['user_id'] = Auth::id();
 
         // creare l'istanza e salvarla nel db
         $post = Post::create($form_data);
@@ -82,8 +89,6 @@ class PostController extends Controller
             // $post->tags()->attach($form_data['tags']);
             $post->tags()->attach($request->tags);
         }
-
-
 
         // redirect alla rotta show
         return to_route('admin.posts.show', $post);
@@ -106,6 +111,11 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+
+        if ($post->user_id !== Auth::id()) {
+            return to_route('admin.posts.index');
+        }
+
         $post->load(['tags']);
         $categories = Category::orderBy('name', 'asc')->get();
         $tags = Tag::orderBy('name', 'asc')->get();
@@ -125,6 +135,9 @@ class PostController extends Controller
         //     'content' => 'nullable|string'
         // ]);
 
+        if ($post->user_id !== Auth::id()) {
+            return to_route('admin.posts.index');
+        }
         // dd($request->all());
 
         $form_data = $request->validated();
@@ -154,7 +167,31 @@ class PostController extends Controller
     {
         // Necessario senza l'onDelete Cascade
         // $post->tags()->detach();
+        if ($post->user_id !== Auth::id()) {
+            return to_route('admin.posts.index');
+        }
+
         $post->delete();
+
+        return to_route('admin.posts.index');
+    }
+
+    public function toggleFavorite(Post $post)
+    {
+
+        // dd($post);
+        $user = Auth::user();
+
+        if ($user->isFavorite($post)) {
+            // se il post è tra i favoriti dell'utente loggato
+            // allora lo rimuoviamo
+            $post->users()->detach($user->id);
+            // $user->favouritePosts()->detach($post->id)
+        } else {
+            $post->users()->attach($user->id);
+        }
+        // altrimenti
+        // aggiungiamo il post tra i preferiti
 
         return to_route('admin.posts.index');
     }
